refactored code after merge
