+++ src/components/SignInForm/props.ts
@@ -1,4 +1,4 @@
-import { IUser } from "redux/reducers/types";
+import { UserType } from "store/slices/auth/auth.type";
 export interface FormValues {
   email: string;
   password: string;
@@ -9,7 +9,7 @@ export interface ILoginFormValues {
 }
 
 export interface IUserData {
-  user: IUser;
+  user: UserType;
   accessToken: string;
   message?: string;
 }
