+++ src/pages/ResetPassword/index.tsx
@@ -20,9 +20,9 @@ import {
 } from "./style";
 import { IPassword } from "./interfaces";
 import { colors } from "constants/index";
-import { useResetPasswordMutation } from "redux/services/passwordApi/passwordApi";
 import ModalWindow from "common/ModalWindow/ModalWindow";
-import { Password } from "redux/models/passwordModels/password.model";
+import { Password } from "store/slices/auth/auth.type";
+import { useResetPasswordMutation } from "store/apis/auth";
 
 const ResetPassword: React.FC = () => {
   const { t } = useTranslation();
@@ -50,7 +50,7 @@ const ResetPassword: React.FC = () => {
       const value: Password = {
         token: params.get("token")?.toString(),
         password: values.confirmPassword,
-        
+
       };
       await resetPassword(value);
       setError(false);
@@ -81,7 +81,7 @@ const ResetPassword: React.FC = () => {
           initialValues={{ remember: true }}
           onFinish={values => onFinish(values as IPassword)}
         >
-       
+
         <StyledSpace>
           <FormItem
             label={t("ResetPage.passwordTitle.item")}
@@ -146,19 +146,19 @@ const ResetPassword: React.FC = () => {
         </FormLink>
       </StyledForm>
       </Wrapper>
-      
+
       <ModalWindow modalIsOpen={modalIsOpen} closeModal={closeModal}>
         <>
-          { 
-            data ? 
-            <WindowTitle level={3}>{t("ResetPage.loginText")}</WindowTitle> 
+          {
+            data ?
+            <WindowTitle level={3}>{t("ResetPage.loginText")}</WindowTitle>
             : dataError
           }
 
           {
-            isError && <WindowTitle level={3}>{t("ResetPage.passwordError")}</WindowTitle> 
+            isError && <WindowTitle level={3}>{t("ResetPage.passwordError")}</WindowTitle>
           }
-        
+
           <NavLink to="/" className="form_link">
             {t("ResetPage.linkText")}
           </NavLink>
