--- src/redux/models/passwordModels/password.model.ts
@@ -1,7 +1,0 @@
-﻿export interface Email {
-  email: string;
-}
-export interface Password {
-  token: string | undefined;
-  password: string;
-}
