+++ src/store/apis/auth/index.ts
@@ -1,4 +1,11 @@
-import { IAuthRequest, IResponse } from "store/slices/auth/auth.type";
+import {
+  Email,
+  IAuthRequest,
+  IResponse,
+  Password,
+} from "store/slices/auth/auth.type";
+//! TODO перенести в типы в соответствующий файл
+import { ILoginFormValues, IUserData } from "components/SignInForm/props";
 import { baseApi } from "..";
 
 export const authApi = baseApi.injectEndpoints({
@@ -28,11 +35,35 @@ export const authApi = baseApi.injectEndpoints({
         },
       }),
     }),
+    login: builder.mutation<IResponse, ILoginFormValues>({
+      query: (body) => ({
+        url: "auth/login",
+        method: "POST",
+        body,
+      }),
+    }),
+    confirmEmail: builder.mutation<Email, string>({
+      query: (email) => ({
+        url: `auth/forgotPassword`,
+        method: "POST",
+        body: email,
+      }),
+    }),
+    resetPassword: builder.mutation<any, Password>({
+      query: (value) => ({
+        url: `auth/changePassword?token=${value.token}`,
+        method: "PUT",
+        body: { password: value.password },
+      }),
+    }),
   }),
 });
 
 export const {
   useCreateUserMutation,
   useLazySignUpUseGoogleQuery,
+  useLoginMutation,
   useLazySelectRoleQuery,
+  useResetPasswordMutation,
+  useConfirmEmailMutation,
 } = authApi;
