+++ src/store/slices/auth/auth.type.ts
@@ -35,5 +35,14 @@ export interface IResponseError  {
 
 type ErrorData = {
   message: string,
-    statusCode: number
+  statusCode: number
 }
+
+export interface Email {
+  email: string;
+}
+export interface Password {
+  token: string | undefined;
+  password: string;
+}
+
